Split Linux metadata runtime modules
